Added exception handling and corrected some small things

+++ Mark5.Mobile.Droid/Utilities/ConversionUtils.cs
@@ -11,7 +11,7 @@ namespace Mark5.Mobile.Droid.Utilities
 {
     public static class ConversionUtils
     {
-        public static int ConvertDpToPixels(int dp)
+        public static int ConvertDpToPixels(float dp)
         {
             return (int)TypedValue.ApplyDimension(ComplexUnitType.Dip, dp, Android.App.Application.Context.Resources.DisplayMetrics);
         }
